Update README.md
# Top100filmes
Welcome to the Repository of my latest and most demanding project, the website Top100filmes. To create this project I dedicated myself to learning some PHP, Laravel, MySQL and Python. I've spent hundreds of hours trying make the best I could with the knowledge I was acquiring through the courses, articles, tutorials, books, etc.
To create this project I also developed an administrative area, to help ease the management of the website, and to help me and my wife on the posting of news, articles, lists, movies, etc to the database which the website uses. The top100filmes is dedicated to helping people find good movies to watch through the top100 sections, and the nominations are based on our personal taste and/or on database filter queries that select the best movies based on the number of votes, users score and it's relevance to the public. I hope you enjoy browsing this repository, and I invite you to access the live website at: https://www.top100filmes.com.br